ldap.events: moved CurrentContainerChangedEvent to new packge
