+++ src/main/java/net/kemitix/ldapmanager/suppliers/CurrentLdapEntityContainerSupplier.java
@@ -24,7 +24,7 @@ SOFTWARE.
 
 package net.kemitix.ldapmanager.suppliers;
 
-import net.kemitix.ldapmanager.events.CurrentContainerChangedEvent;
+import net.kemitix.ldapmanager.ldap.events.CurrentContainerChangedEvent;
 import net.kemitix.ldapmanager.state.CurrentContainer;
 import net.kemitix.ldapmanager.state.LdapEntityContainer;
 import net.kemitix.ldapmanager.state.LdapEntityContainerMap;
