+++ src/test/java/net/kemitix/ldapmanager/state/DefaultCurrentContainerTest.java
@@ -2,7 +2,7 @@ package net.kemitix.ldapmanager.state;
 
 import lombok.val;
 import net.kemitix.ldapmanager.domain.OU;
-import net.kemitix.ldapmanager.events.CurrentContainerChangedEvent;
+import net.kemitix.ldapmanager.ldap.events.CurrentContainerChangedEvent;
 import net.kemitix.ldapmanager.ldap.LdapNameUtil;
 import net.kemitix.ldapmanager.navigation.events.NavigationItemOuActionEvent;
 import org.junit.Before;
